fix(observe): preserve clickhouse lexical regions
