+++ crates/temper-observe/src/clickhouse.rs
@@ -27,8 +27,14 @@ enum QueryRegion {
     SingleQuoted,
     DoubleQuoted,
     BacktickQuoted,
+    UnicodeSingleQuoted,
+    UnicodeDoubleQuoted,
     LineComment,
-    BlockComment,
+    BlockComment(usize),
+    Heredoc {
+        delimiter_start: usize,
+        delimiter_len: usize,
+    },
 }
 
 impl ClickHouseStore {
@@ -71,13 +77,52 @@ impl ClickHouseStore {
 
             match region {
                 QueryRegion::Code => {
+                    if c == b'$'
+                        && let Some(delimiter_len) = heredoc_delimiter_len(sql, i)
+                    {
+                        out.push_str(&sql[i..i + delimiter_len]);
+                        let delimiter_start = i;
+                        i += delimiter_len;
+                        region = QueryRegion::Heredoc {
+                            delimiter_start,
+                            delimiter_len,
+                        };
+                        continue;
+                    }
+                    if is_clickhouse_word_char(c) {
+                        let mut end = i + 1;
+                        while end < bytes.len()
+                            && (is_clickhouse_word_char(bytes[end]) || bytes[end] == b'$')
+                        {
+                            end += 1;
+                        }
+                        out.push_str(&sql[i..end]);
+                        i = end;
+                        continue;
+                    }
                     let next = bytes.get(i + 1).copied();
                     region = match (c, next) {
                         (b'\'', _) => QueryRegion::SingleQuoted,
                         (b'"', _) => QueryRegion::DoubleQuoted,
                         (b'`', _) => QueryRegion::BacktickQuoted,
-                        (b'-', Some(b'-')) | (b'#', _) => QueryRegion::LineComment,
-                        (b'/', Some(b'*')) => QueryRegion::BlockComment,
+                        (b'-', Some(b'-')) | (b'/', Some(b'/')) => QueryRegion::LineComment,
+                        (b'#', Some(b' ' | b'!')) => QueryRegion::LineComment,
+                        (b'/', Some(b'*')) => {
+                            out.push_str("/*");
+                            i += 2;
+                            region = QueryRegion::BlockComment(1);
+                            continue;
+                        }
+                        (0xE2, Some(0x80)) if bytes.get(i + 2) == Some(&0x98) => {
+                            push_char(sql, &mut out, &mut i);
+                            region = QueryRegion::UnicodeSingleQuoted;
+                            continue;
+                        }
+                        (0xE2, Some(0x80)) if bytes.get(i + 2) == Some(&0x9C) => {
+                            push_char(sql, &mut out, &mut i);
+                            region = QueryRegion::UnicodeDoubleQuoted;
+                            continue;
+                        }
                         (b'$', Some(digit)) if digit.is_ascii_digit() => {
                             let mut end = i + 1;
                             while end < bytes.len() && bytes[end].is_ascii_digit() {
@@ -143,10 +188,42 @@ impl ClickHouseStore {
                         region = QueryRegion::Code;
                     }
                 }
-                QueryRegion::BlockComment => {
+                QueryRegion::BlockComment(depth) => {
+                    if c == b'/' && bytes.get(i + 1) == Some(&b'*') {
+                        out.push_str("/*");
+                        i += 2;
+                        region = QueryRegion::BlockComment(depth + 1);
+                        continue;
+                    }
                     if c == b'*' && bytes.get(i + 1) == Some(&b'/') {
                         out.push_str("*/");
                         i += 2;
+                        region = if depth == 1 {
+                            QueryRegion::Code
+                        } else {
+                            QueryRegion::BlockComment(depth - 1)
+                        };
+                        continue;
+                    }
+                }
+                QueryRegion::UnicodeSingleQuoted | QueryRegion::UnicodeDoubleQuoted => {
+                    let closing = if region == QueryRegion::UnicodeSingleQuoted {
+                        "’"
+                    } else {
+                        "”"
+                    };
+                    if sql[i..].starts_with(closing) {
+                        region = QueryRegion::Code;
+                    }
+                }
+                QueryRegion::Heredoc {
+                    delimiter_start,
+                    delimiter_len,
+                } => {
+                    let delimiter = &sql[delimiter_start..delimiter_start + delimiter_len];
+                    if sql[i..].starts_with(delimiter) {
+                        out.push_str(delimiter);
+                        i += delimiter_len;
                         region = QueryRegion::Code;
                         continue;
                     }
@@ -249,6 +326,27 @@ fn clickhouse_param(param: &SqlParam) -> Option<(&'static str, String)> {
     }
 }
 
+fn heredoc_delimiter_len(source: &str, start: usize) -> Option<usize> {
+    let bytes = source.as_bytes();
+    let tag_len = bytes[start + 1..].iter().position(|byte| *byte == b'$')?;
+    let delimiter_len = tag_len + 2;
+    let tag = &bytes[start + 1..start + 1 + tag_len];
+    if !tag
+        .iter()
+        .all(|byte| byte.is_ascii_alphanumeric() || *byte == b'_')
+    {
+        return None;
+    }
+    let delimiter = &source[start..start + delimiter_len];
+    source[start + delimiter_len..]
+        .contains(delimiter)
+        .then_some(delimiter_len)
+}
+
+fn is_clickhouse_word_char(byte: u8) -> bool {
+    byte.is_ascii_alphanumeric() || byte == b'_'
+}
+
 fn push_char(source: &str, output: &mut String, index: &mut usize) {
     let ch = source[*index..]
         .chars()
@@ -294,151 +392,5 @@ fn parse_json_each_row(body: &str) -> Result<ResultSet, ObserveError> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_store_construction() {
-        let store = ClickHouseStore::new("http://localhost:8123");
-        assert_eq!(store.base_url(), "http://localhost:8123");
-    }
-
-    #[test]
-    fn bind_params_uses_clickhouse_typed_placeholders() {
-        let sql = "SELECT * FROM spans WHERE service = $1 AND duration_ns > $2";
-        let params = vec![SqlParam::String("api".into()), SqlParam::Int(1000)];
-        let result = ClickHouseStore::bind_params(sql, &params).unwrap();
-        assert_eq!(
-            result,
-            BoundClickHouseQuery {
-                sql: "SELECT * FROM spans WHERE service = {p1:String} AND duration_ns > {p2:Int64}"
-                    .into(),
-                form_params: vec![
-                    ("param_p1".into(), "api".into()),
-                    ("param_p2".into(), "1000".into()),
-                ],
-            }
-        );
-    }
-
-    #[test]
-    fn bind_params_preserves_placeholders_in_quoted_regions_and_comments() {
-        let sql = "SELECT '$1', \"$1\", `$1`, $1 -- $1\n/* $1 */ # $1\n";
-        let params = vec![SqlParam::String("api".into())];
-        let result = ClickHouseStore::bind_params(sql, &params).unwrap();
-        assert_eq!(
-            result.sql,
-            "SELECT '$1', \"$1\", `$1`, {p1:String} -- $1\n/* $1 */ # $1\n"
-        );
-    }
-
-    #[test]
-    fn bind_params_never_places_attacker_text_in_sql() {
-        let attack = "x\\' OR 1=1; DROP TABLE spans; -- $2";
-        let result = ClickHouseStore::bind_params(
-            "SELECT * FROM spans WHERE a = $1 AND b = $2",
-            &[
-                SqlParam::String(attack.into()),
-                SqlParam::String("second".into()),
-            ],
-        )
-        .unwrap();
-
-        assert_eq!(
-            result.sql,
-            "SELECT * FROM spans WHERE a = {p1:String} AND b = {p2:String}"
-        );
-        assert!(!result.sql.contains("DROP TABLE"));
-        assert_eq!(result.form_params[0].1, attack);
-    }
-
-    #[test]
-    fn bind_params_handles_types_two_digits_and_multibyte_text() {
-        let sql = "SELECT $10, name = $1, note = 'café ☕'";
-        let mut params: Vec<SqlParam> = (1..=10).map(SqlParam::Int).collect();
-        params[0] = SqlParam::String("first".into());
-        let result = ClickHouseStore::bind_params(sql, &params).unwrap();
-        assert_eq!(
-            result.sql,
-            "SELECT {p10:Int64}, name = {p1:String}, note = 'café ☕'"
-        );
-        assert_eq!(
-            result.form_params,
-            vec![
-                ("param_p1".into(), "first".into()),
-                ("param_p10".into(), "10".into()),
-            ]
-        );
-    }
-
-    #[test]
-    fn bind_params_handles_null_bool_and_float() {
-        let result = ClickHouseStore::bind_params(
-            "SELECT $1, $2, $3",
-            &[SqlParam::Null, SqlParam::Bool(true), SqlParam::Float(1.5)],
-        )
-        .unwrap();
-        assert_eq!(result.sql, "SELECT NULL, {p2:UInt8}, {p3:Float64}");
-        assert_eq!(
-            result.form_params,
-            vec![
-                ("param_p2".into(), "1".into()),
-                ("param_p3".into(), "1.5".into()),
-            ]
-        );
-    }
-
-    #[test]
-    fn bind_params_rejects_missing_and_zero_parameters() {
-        let missing = ClickHouseStore::bind_params("SELECT $2", &[SqlParam::Int(1)]);
-        assert!(matches!(missing, Err(ObserveError::InvalidQuery(_))));
-
-        let zero = ClickHouseStore::bind_params("SELECT $0", &[SqlParam::Int(1)]);
-        assert!(matches!(zero, Err(ObserveError::InvalidQuery(_))));
-    }
-
-    #[test]
-    fn test_parse_empty() {
-        let rs = parse_json_each_row("").unwrap();
-        assert!(rs.is_empty());
-    }
-
-    #[test]
-    fn test_parse_single_row() {
-        let body = r#"{"service":"api","status":"ok"}"#;
-        let rs = parse_json_each_row(body).unwrap();
-        assert_eq!(rs.len(), 1);
-        assert_eq!(rs.get(0, "service"), Some(&serde_json::json!("api")));
-    }
-
-    #[test]
-    fn test_parse_multiple_rows() {
-        let body = "{\"a\":1}\n{\"a\":2}";
-        let rs = parse_json_each_row(body).unwrap();
-        assert_eq!(rs.len(), 2);
-    }
-
-    #[test]
-    fn build_request_keeps_parameters_out_of_url() {
-        let attack = "x\\' OR 1=1 --";
-        let store = ClickHouseStore::new("http://127.0.0.1:8123");
-        let request = store
-            .build_request(
-                "SELECT service FROM spans WHERE service = $1",
-                &[SqlParam::String(attack.into())],
-            )
-            .unwrap();
-
-        assert_eq!(request.method(), reqwest::Method::POST);
-        assert_eq!(request.url().path(), "/");
-        assert_eq!(request.url().query(), Some("default_format=JSONEachRow"));
-        assert!(!request.url().as_str().contains(attack));
-        assert!(
-            request
-                .headers()
-                .get(reqwest::header::CONTENT_TYPE)
-                .and_then(|value| value.to_str().ok())
-                .is_some_and(|value| value.starts_with("multipart/form-data; boundary="))
-        );
-    }
-}
+#[path = "clickhouse_tests.rs"]
+mod tests;
